Pre-production release
+ All functions operational
+ Developer functions active (double tap recent contacts for menu)
+ Reliable continuous tracking of iOS device from initial contact

Limitations due to iOS BLE UUID change
- Background iOS device tracking is only possible from first contact with the device. If a device goes out of range for circa 20 minutes and return back to range, it will not be detected again until the app on one of the devices is brought back to foreground mode.
- Background tracking can handle Airplane mode cycle for under one minute (e.g. accidental user action), but is likely to behave the same way as going out of range for extended periods, i.e. the device won't be detected again for tracking until the app on one of the devices is brought back to foreground mode.

The current limitations are only applicable to iOS devices operating in background mode (expected norm). This is caused by iOS changing its peripheral identifier (UUID) following a period of inactivity or airplane mode ON/OFF cycle, as such another iOS device acting as central with a pending connect to this peripheral is now using an expired UUID for connection, and the connect is not satisfied even when the peripheral comes back to range. Unfortunately, the didDiscover delegate method is also not called despite the change in UUID, thus the central is no mechanism for reconnecting nor connecting to the peripheral again.

Ideas that have not fully resolve this situation:
1. Using pending write to CBPeripheralManager to obtain the updated UUID. This has enabled survival of Airplane mode ON/OFF cycle for short periods (1 minute), when the UUID has indeed changed.
2. Using pending setNotifyValue in CBCentralManager to reconnect based on new UUID. This also enables survival of Airplane mode ON/OFF cycle for short periods (1 minute), when the UUID has changed.

Ideas that have made no difference:
1. Instantiating a new CentralManager to scanForPeripheral, made no difference, the second instance does not receive a didDiscover call, if the peripheral has already been discovered by the first instance.
2. Changing service UUID being advertised by PeripheralManager and service UUIDs being scanned by CentralManager, made no difference, didDiscover is not called.